fix: avoid breaking older subapp-server modules (#1663)
