+++ packages/subapp-server/lib/utils.js
@@ -8,7 +8,7 @@ const Fs = require("fs");
 const Path = require("path");
 const Boom = require("@hapi/boom");
 const optionalRequire = require("optional-require")(require);
-const getDevProxy = optionalRequire("@xarc/app-dev/config/dev-proxy");
+const getDevProxy = optionalRequire("@xarc/app-dev/config/get-dev-proxy");
 const HttpStatusCodes = require("http-status-codes");
 
 /**
@@ -62,12 +62,9 @@ const updateFullTemplate = (baseDir, options) => {
 };
 
 function getDefaultRouteOptions() {
-  const {
-    settings = {},
-    devServer = {},
-    fullDevServer = {},
-    httpDevServer = {}
-  } = getDevProxy ? getDevProxy() : {};
+  const { settings = {}, devServer = {}, fullDevServer = {}, httpDevServer = {} } = getDevProxy
+    ? getDevProxy()
+    : {};
   const { webpackDev, useDevProxy } = settings;
   // temporary location to write build artifacts in dev mode
   const buildArtifacts = ".etmp";
