+++ packages/xarc-app-dev/lib/dev-admin/admin-server.js
@@ -24,7 +24,7 @@ const {
   settings: { useDevProxy: DEV_PROXY_ENABLED, adminLogLevel },
   fullDevServer,
   controlPaths
-} = require("../../config/dev-proxy")();
+} = require("../../config/dev-proxy");
 
 const ADMIN_LOG_LEVEL = parseInt(adminLogLevel) || 0;
